@@ -5,6 +5,7 @@
   "author": "Benedikt Deicke <benedikt@userlist.com>",
   "homepage": "https://github.com/userlist/userlist-javascript",
   "license": "MIT",
+  "type": "module",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "files": [
@@ -23,7 +24,7 @@
   "scripts": {
     "lint": "eslint src test",
     "build": "rollup -c",
-    "test": "mocha --require test/helper --require @babel/register --recursive --colors",
+    "test": "mocha",
     "test:watch": "pnpm test --watch",
     "prepublish": "pnpm lint && pnpm test && pnpm build"
   },
@@ -32,29 +33,32 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@babel/eslint-parser": "^7.26.10",
     "@babel/plugin-transform-runtime": "^7.9.0",
     "@babel/preset-env": "^7.9.0",
     "@babel/register": "^7.9.0",
     "@babel/runtime": "^7.9.2",
-    "@rollup/plugin-commonjs": "^11.0.2",
-    "@rollup/plugin-node-resolve": "^7.1.1",
-    "babel-eslint": "^10.1.0",
-    "chai": "^4.2.0",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.1",
+    "@rollup/plugin-babel": "^6.0.0",
+    "@rollup/plugin-commonjs": "^28.0.0",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "chai": "^5.0.0",
+    "eslint": "^9.0.0",
+    "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "jsdom": "^16.2.1",
-    "mocha": "^7.1.1",
-    "rollup": "^2.2.0",
-    "rollup-plugin-babel": "^4.4.0",
-    "sinon": "^9.0.1",
-    "sinon-chai": "^3.5.0",
-    "ws": "^7.2.3"
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-promise": "^7.0.0",
+    "globals": "^16.0.0",
+    "jsdom": "^26.0.0",
+    "mocha": "^11.0.0",
+    "prettier": "^3.5.3",
+    "rollup": "^4.0.0",
+    "sinon": "^19.0.0",
+    "sinon-chai": "^4.0.0",
+    "ws": "^8.0.0"
   },
   "dependencies": {
+    "@eslint/js": "^9.22.0",
     "@rails/actioncable": "^6.0.2-2",
     "@userlist/widget": "^0.3.8"
   },
