@@ -1,23 +1,23 @@
-var { JSDOM } = require('jsdom');
+import { JSDOM } from 'jsdom';
 
 var { window } = new JSDOM(`<!DOCTYPE html><html><body></body></html>`);
 
-if(global.window === undefined) {
+if (global.window === undefined) {
   global.window = window;
 }
 
-if(global.document === undefined) {
+if (global.document === undefined) {
   global.document = window.document;
 }
 
-if(global.self === undefined) {
+if (global.self === undefined) {
   global.self = window;
 }
 
-if(global.addEventListener === undefined) {
+if (global.addEventListener === undefined) {
   global.addEventListener = window.addEventListener;
 }
 
-if(global.removeEventListener === undefined) {
+if (global.removeEventListener === undefined) {
   global.removeEventListener = window.removeEventListener;
 }
