@@ -1,7 +1,8 @@
-import babel from 'rollup-plugin-babel';
+import babel from '@rollup/plugin-babel';
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import pkg from './package.json';
+
+import pkg from './package.json' with { type: "json" };
 
 const plugins = [
   resolve({
@@ -13,7 +14,7 @@ const plugins = [
     }
   }),
   babel({
-    runtimeHelpers: true
+    babelHelpers: 'runtime'
   })
 ];
 
